+++ src/main/java/com/techbow/homework/y2021/m09/Jacob/LC_746.java
@@ -9,7 +9,7 @@ public class LC_746 {
         dp[0] = cost[0];
         dp[1] = cost[1];
         for (int i = 2; i < cost.length; i++) {
-            dp[i] = Math.min(dp[i-1], dp[i-2] + cost(i));
+            dp[i] = Math.min(dp[i-1], dp[i-2] + cost[i]);
         }
         return Math.min(dp[n-1], dp[n-2]);
     }
