+++ src/main/java/com/techbow/homework/y2021/m09/WooWooLily/LC_160.java
@@ -1,5 +1,13 @@
-package com.techbow.homework.y2021.m09.WooWooLily;
+package com.techbow.homework.y2021.m09.WooWooLily;//package com.techbow.homework.y2021.m09.WooWooLily;
+
 import java.util.*;
+class ListNode {
+    int val;
+    ListNode next;
+    ListNode() {}
+    ListNode(int val) { this.val = val; }
+    ListNode(int val, ListNode next) { this.val = val; this.next = next; }
+}
 // length headA: a + c
 // length headB: b + c
 // if we make headA move what headB has moved and vice versa, a+c+b+c vs b+c+a+c, these two ptr will move together
