+++ src/main/java/com/techbow/homework/y2021/m09/boyuan/小宋小宋/BFS_102.java
@@ -1,22 +1,11 @@
 package com.techbow.homework.y2021.m09.boyuan.小宋小宋;
 
-import javax.swing.tree.TreeNode;
+
 import java.util.ArrayList;
 import java.util.List;
 
 
-public class TreeNode {
-    int val;
-     TreeNode left;
-    TreeNode right;
-     TreeNode() {}
-    TreeNode(int val) { this.val = val; }
-    TreeNode(int val, TreeNode left, TreeNode right) {
-         this.val = val;
-         this.left = left;
-          this.right = right;
-      }
- }
+
 public class BFS_102 {
     List<List<Integer>> res = new ArrayList<>();
     public List<List<Integer>> levelOrder(TreeNode root) {
