+++ src/main/java/com/techbow/homework/y2021/m09/boyuan/小宋小宋/TreeNode.java
@@ -1,0 +1,16 @@
+package com.techbow.homework.y2021.m09.boyuan.小宋小宋;
+
+
+public class TreeNode {
+    int val;
+     TreeNode left;
+    TreeNode right;
+     TreeNode() {}
+    TreeNode(int val) { this.val = val; }
+    TreeNode(int val, TreeNode left, TreeNode right) {
+         this.val = val;
+         this.left = left;
+          this.right = right;
+      }
+ }
+
