--- src/main/java/com/techbow/homework/y2021/m09/boyuan/小宋小宋/TreeNodeImpl.java
@@ -1,4 +1,0 @@
-package com.techbow.homework.y2021.m09.boyuan.小宋小宋;
-
-public class TreeNodeImpl extends TreeNode {
-}
