+++ src/main/java/com/techbow/homework/y2021/m09/ludai/MaximumEarningTaxi.java
@@ -31,4 +31,4 @@ public class MaximumEarningTaxi {
             return dp.get(left)[0]<= t? left: left-1;
         }
     }
-}
+
