+++ src/main/java/com/techbow/homework/y2021/m09/ludai/Maximumearining.java
@@ -32,4 +32,4 @@ public class Maximumearining {
         return dp.get(left)[0]<= t? left: left-1;
     }
 }
-}
+
